refactor main, router and handler
